Introduce LoxType enum
I need a way of having multiple types in the "literal" field of a Token.
I think this is what the original Java implementation is using Object
for.

Doing dynamism is hard in Rust. I tried to implement this as a trait,
that I could implement on all the underlying types I cared about. so I
could push in a String, or an i32 directly or something like that, but I
couldn't make it work in teh time I allowed myself so I an using an enum
instead. I actually like the explicitness of it, even if it does mean I
have to wrap everything in LoxType::Type.

I hope this works later on in the book