Uppercase the TokenTypes
This requires disabling the non_camel_case_type_names linter warning,
but the payoff is that I don't have to do any string manipulation in the
Display trait impl for the Token in order to match teh output that the
craftinginterpreter test suite is expecting.

Because of this I can remove my only dependency so far.